refactor: some hooks name changed (#1821)
* refactor: rename useStakingRewards3.ts to useStakingRewardsChart.ts

* refactor: remove useIdentity.ts and use useIdentity2.ts

* refactor: rename useStakingRewards2.ts to useSoloStakingTotalReward.ts

* refactor: rename useTransactionHistory2 to useTransactionHistory
