+++ packages/extension-polkagate/src/hooks/index.ts
@@ -45,7 +45,6 @@ export { default as useGenesisHashOptions } from './useGenesisHashOptions';
 export { default as useHasDelegated } from './useHasDelegated';
 export { default as useHasIdentityTooltipText } from './useHasIdentityTooltipText';
 export { default as useIdentity } from './useIdentity';
-export { default as useIdentity2 } from './useIdentity2';
 export { default as useIsAutoEndpoint } from './useIsAutoEndpoint';
 export { default as useIsBlueish } from './useIsBlueish';
 export { default as useIsDark } from './useIsDark';
@@ -90,13 +89,13 @@ export { default as useSelectedChains } from './useSelectedChains';
 export { default as useSelectedLanguage } from './useSelectedLanguage';
 export { default as useSelectedProfile } from './useSelectedProfile';
 export { default as useSoloStakingInfo } from './useSoloStakingInfo';
+export { default as useSoloStakingTotalReward } from './useSoloStakingTotalReward';
 export { default as useStakingAccount } from './useStakingAccount';
 export * from './useStakingApi';
 export { default as useStakingConsts } from './useStakingConsts';
 export { default as useStakingPositions } from './useStakingPositions';
 export { default as useStakingRewardDestinationAddress } from './useStakingRewardDestinationAddress';
-export { default as useStakingRewards2 } from './useStakingRewards2';
-export { default as useStakingRewards3 } from './useStakingRewards3';
+export { default as useStakingRewardsChart } from './useStakingRewardsChart';
 export { default as useStashId } from './useStashId';
 export { default as useTeleport } from './useTeleport';
 export { default as useTimeToUnlock } from './useTimeToUnlock';
