+++ packages/extension-polkagate/src/hooks/usePendingRewards.ts
@@ -81,7 +81,7 @@ const isRewardsPaged = (chainName: string | undefined, era: number): boolean => 
 
 export const MAX_SUPPORTED_PAYOUT_ERAS = 7; // TODO: can increase to more if needed after enough tests
 
-export default function usePendingRewards3 (address: string | undefined, genesisHash: string | undefined): UnclaimedPayouts | undefined {
+export default function usePendingRewards (address: string | undefined, genesisHash: string | undefined): UnclaimedPayouts | undefined {
   const { api, chainName } = useChainInfo(genesisHash);
   const formatted = useFormatted(address, genesisHash);
   const activeEra = useActiveEraIndex(genesisHash);
