+++ packages/extension-polkagate/src/hooks/useStakingRewardsChart.ts
@@ -89,7 +89,7 @@ export interface UseStakingRewards {
  * @param isFullScreen - Whether the chart is displayed in full screen mode
  * @returns Object containing chart data, options, and control functions
  */
-export default function useStakingRewards3 (address: string | undefined, genesisHash: string | undefined, type: 'solo' | 'pool', isFullScreen?: boolean): UseStakingRewards {
+export default function useStakingRewardsChart (address: string | undefined, genesisHash: string | undefined, type: 'solo' | 'pool', isFullScreen?: boolean): UseStakingRewards {
   const theme = useTheme();
   const { chainName, decimal, token } = useChainInfo(genesisHash, true);
