+++ packages/extension-polkagate/src/popup/staking/partial/StakingRewardTile.tsx
@@ -13,7 +13,7 @@ import { BN_ZERO } from '@polkadot/util';
 
 import { Thunder } from '../../../assets/gif';
 import { FormatBalance2, FormatPrice, MySkeleton } from '../../../components';
-import { useChainInfo, usePrices, useStakingRewards3, useTokenPrice, useTranslation } from '../../../hooks';
+import { useChainInfo, usePrices, useStakingRewardsChart, useTokenPrice, useTranslation } from '../../../hooks';
 import { Background } from '../../../style';
 import { ColumnAmounts } from '../../tokens/partial/ColumnAmounts';
 import StakingActionButton from './StakingActionButton';
@@ -252,7 +252,7 @@ export default function StakingRewardTile ({ address, genesisHash, isDisabled, l
   const pricesInCurrency = usePrices();
   const tokenPrice = useTokenPrice(genesisHash);
   // Pool total earned rewards
-  const { totalClaimedReward } = useStakingRewards3(address, genesisHash, type);
+  const { totalClaimedReward } = useStakingRewardsChart(address, genesisHash, type);
 
   const rewardInCurrency = useMemo(() => {
     if (!reward || !pricesInCurrency || !tokenPrice || !decimal) {
